kaggle_package/requirements.txt
# Skeleton-Based Rehabilitation Exercise Quality Grading System
# Python 3.8+

# Deep Learning
tensorflow>=2.10.0
keras>=2.10.0

# Pose & Video
mediapipe>=0.10.0
opencv-python>=4.8.0

# Data & Utils
numpy>=1.23.0
pandas>=1.5.0
scipy>=1.10.0

# UI
pygame>=2.5.0

# Audio feedback (TTS for live feedback)
pyttsx3>=2.90

# Config & Logging
pyyaml>=6.0
tqdm>=4.65.0

# Optional: Visualization & Logging
matplotlib>=3.7.0
tensorboard>=2.12.0

# Optional: Download training/demo videos from YouTube (scripts/download_training_videos.py)
# yt-dlp>=2024.1.0
